Handle return from edit playlist to plugin entry. Simplify screen action creation

+++ dune_plugin/lib/dune_default_sqlite_engine.php
@@ -786,8 +786,10 @@ class Dune_Default_Sqlite_Engine
      */
     public function remove_cookie($name)
     {
-        $table_name = self::COOKIES_TABLE;
-        $this->sql_playlist->exec("DELETE FROM $table_name WHERE param = '$name';");
+        if ($this->sql_playlist) {
+            $table_name = self::COOKIES_TABLE;
+            $this->sql_playlist->exec("DELETE FROM $table_name WHERE param = '$name';");
+        }
     }
 
     ///////////////////////////////////////////////////////////////////////
